+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# [next]
+# [2.7.6]
 - Adds mixin `Acceleration`.
 - Rename methods in `MouseGesture`
   - `onHoverScreen` to `onMouseHoverScreen`